update: split gateway.py
- split the gateway file into controller, routes & schema
- changed main to new import
- deleted gateway.py
- added api_routes.py to congregate the logic
- added init files
- added pathfinding_schema.py that defines the pathfindingrequest schema
- added pathfinding_service.py that contains the logic for the fastest_path endpoint
- changed tests to reflect new file structure
