+++ app/schemas/pathfinding_schema.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+class PathfindingRequest(BaseModel):
+    source: str
+    target: str
